+++ src/main/java/io/github/devatherock/artifactory/util/BadgeGenerator.java
@@ -1,10 +1,11 @@
 package io.github.devatherock.artifactory.util;
 
-import io.github.devatherock.artifactory.config.ArtifactoryBadgeProperties;
+import io.github.devatherock.artifactory.config.ShieldsIOProperties;
 import io.github.devatherock.artifactory.service.ShieldsIOClient;
 import io.micronaut.core.io.IOUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
@@ -16,25 +17,26 @@ import java.io.InputStreamReader;
 import java.net.URI;
 import java.net.URL;
 
+@Slf4j
 @Singleton
 @RequiredArgsConstructor
 public class BadgeGenerator {
-    private static final float WIDTH_ONE_CHAR = 5.8f;
-    private static final float TEXT_LENGTH_MULTIPLIER = 8.5f;
+    private static final int WIDTH_ONE_CHAR = 8;
+
     private final ShieldsIOClient shieldsIOClient;
     private final HttpClient httpClient;
-    private final ArtifactoryBadgeProperties config;
+    private final ShieldsIOProperties config;
 
     @SneakyThrows
     public String generateBadge(String label, String value) {
         String badge = null;
 
-        if (config.getShieldsIo().isEnabled()) {
+        if (config.isEnabled()) {
             String shieldIoImageName = String.format("%s-%s-blue.svg", label, value);
 
-            if (config.getShieldsIo().isCustomHttp()) {
+            if (config.isCustomHttp()) {
                 // For URL encoding spaces in the path. URLEncoder.encode encodes them into + instead of %20
-                URL url = new URL(config.getShieldsIo().getUrl() + "/" + shieldIoImageName);
+                URL url = new URL(config.getUrl() + "/" + shieldIoImageName);
                 URI uri = new URI(url.getProtocol(), url.getUserInfo(), url.getHost(), url.getPort(), url.getPath(),
                         url.getQuery(), url.getRef());
                 HttpGet httpGet = new HttpGet(uri);
@@ -55,21 +57,13 @@ public class BadgeGenerator {
     }
 
     private String generateSvg(String label, String value) {
-        float labelWidthFloat = (label.length() + 2) * WIDTH_ONE_CHAR;
-        float valueWidthFloat = (value.length() + 2) * WIDTH_ONE_CHAR;
-        float labelLengthFloat = labelWidthFloat * TEXT_LENGTH_MULTIPLIER;
-        float valueLengthFloat = valueWidthFloat * TEXT_LENGTH_MULTIPLIER;
-
-        String title = label + ": " + value;
-        int labelWidth = (int) (labelWidthFloat);
-        int valueWidth = (int) (valueWidthFloat);
+        LOGGER.debug("Generating custom SVG");
+        int labelWidth = (label.length() + 2) * WIDTH_ONE_CHAR;
+        int valueWidth = (value.length() + 2) * WIDTH_ONE_CHAR;
         int totalWidth = labelWidth + valueWidth;
-        int labelLength = (int) labelLengthFloat;
-        int valueLength = (int) valueLengthFloat;
-        int labelStartPosition = (int) ((labelLengthFloat / 2) + (WIDTH_ONE_CHAR * TEXT_LENGTH_MULTIPLIER));
-        int valueStartPosition = (int) (labelLengthFloat + (valueLengthFloat * 1.7 / 2));
-        StringBuilder svg = new StringBuilder();
+        String title = label + ": " + value;
 
+        StringBuilder svg = new StringBuilder();
         svg.append(
                 String.format("<svg xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" width=\"%d\" height=\"20\" role=\"img\" aria-label=\"%s\">",
                         totalWidth, title));
@@ -89,17 +83,12 @@ public class BadgeGenerator {
                 labelWidth, valueWidth));
         svg.append(String.format("<rect width=\"%d\" height=\"20\" fill=\"url(#s)\"/>", totalWidth));
         svg.append("</g>");
-        svg.append("<g fill=\"#fff\" text-anchor=\"middle\" font-family=\"Verdana,Geneva,DejaVu Sans,sans-serif\" text-rendering=\"geometricPrecision\" font-size=\"110\">");
+        svg.append("<g font-family=\"mono\" font-size=\"smaller\">");
         svg.append(
-                String.format("<text aria-hidden=\"true\" x=\"%d\" y=\"150\" fill=\"#010101\" fill-opacity=\".3\" transform=\"scale(.1)\" textLength=\"%d\">%s</text>",
-                        labelStartPosition, labelLength, label));
-        svg.append(String.format("<text x=\"%d\" y=\"140\" transform=\"scale(.1)\" fill=\"#fff\" textLength=\"%d\">%s</text>",
-                labelStartPosition, labelLength, label));
-        svg.append(
-                String.format("<text aria-hidden=\"true\" x=\"%d\" y=\"150\" fill=\"#010101\" fill-opacity=\".3\" transform=\"scale(.1)\" textLength=\"%d\">%s</text>",
-                        valueStartPosition, valueLength, value));
-        svg.append(String.format("<text x=\"%d\" y=\"140\" transform=\"scale(.1)\" fill=\"#fff\" textLength=\"%d\">%s</text>",
-                valueStartPosition, valueLength, value));
+                String.format("<text aria-hidden=\"true\" x=\"0\" y=\"15\" fill=\"#fff\" xml:space=\"preserve\"> %s </text>",
+                        label));
+        svg.append(String.format("<text aria-hidden=\"true\" x=\"%d\" y=\"15\" fill=\"#fff\" xml:space=\"preserve\"> %s </text>",
+                labelWidth, value));
         svg.append("</g>");
         svg.append("</svg>");