+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Initial version that generates docker pulls, image size and layers badges
 - `/health` endpoint
+- Custom badge generator
 
 ### Changed
 - Caught exception when JSON processing fails